Fix compiler error when using -Wconversion.
/mariadb/10.3/wsrep/wsrep_gtid.c:45:26: error: conversion from 'long int' to 'unsigned int' may change value [-Werror=conversion]
   45 |                 offset = endptr - str;

+++ wsrep/wsrep_gtid.c
@@ -42,7 +42,7 @@ wsrep_gtid_scan(const char* str, size_t str_len, wsrep_gtid_t* gtid)
             gtid->seqno = strtoll(str + offset, &endptr, 0);
 
             if (errno == 0) {
-                offset = endptr - str;
+                offset = (unsigned int)(endptr - str);
                 return offset;
             }
         }
